Move blueprint to a factory.
